+++ slack/sentry/versions.tf
@@ -1,0 +1,9 @@
+terraform {
+  required_providers {
+    sentry = {
+      source  = "jianyuan/sentry"
+      version = ">= 0.11"
+    }
+  }
+  required_version = ">= 1.0"
+}