+++ stack/app/rds.tf
@@ -34,15 +34,18 @@ module "rds" {
   ]
 
   # optional
-  username               = var.rds_is_read_replica ? null : local.credentials.db_username
-  instance_class         = var.rds_instance_class
-  engine_version         = var.rds_engine_version
-  allocated_storage      = var.rds_allocated_storage
-  multi_az               = var.rds_multi_az == null ? var.environment == "production" : var.rds_multi_az
-  master_db_instance_arn = var.rds_master_db_instance_arn
-  is_read_replica        = var.rds_is_read_replica
-  regional               = var.regional
-  name                   = var.rds_name
-  allow_from_cidr_blocks = var.rds_allow_from_cidr_blocks
-  subnet_group_name      = var.rds_subnet_group_name
+  username                  = var.rds_is_read_replica ? null : local.credentials.db_username
+  instance_class            = var.rds_instance_class
+  engine_version            = var.rds_engine_version
+  allocated_storage         = var.rds_allocated_storage
+  multi_az                  = var.rds_multi_az == null ? var.environment == "production" : var.rds_multi_az
+  master_db_instance_arn    = var.rds_master_db_instance_arn
+  is_read_replica           = var.rds_is_read_replica
+  regional                  = var.regional
+  name                      = var.rds_name
+  allow_from_cidr_blocks    = var.rds_allow_from_cidr_blocks
+  subnet_group_name         = var.rds_subnet_group_name
+  delete_automated_backups  = var.rds_delete_automated_backups
+  skip_final_snapshot       = var.rds_skip_final_snapshot
+  final_snapshot_identifier = var.rds_final_snapshot_identifier
 }